+++ internal/testing/testdata/exampledata/alpine-cyclonedx.json
