+++ internal/testing/testdata/exampledata/alpine-small-spdx.json
