+++ internal/testing/testdata/exampledata/alpine-spdx.json
