+++ internal/testing/testdata/exampledata/big-mongo-cyclonedx.json
