+++ internal/testing/testdata/exampledata/busybox-cyclonedx.json
