+++ internal/testing/testdata/exampledata/distroless-cyclonedx.json
