+++ internal/testing/testdata/exampledata/invalid-cyclonedx.json
