+++ internal/testing/testdata/exampledata/invalid-spdx.json
