+++ internal/testing/testdata/exampledata/kubernetes-scorecard.json
