+++ internal/testing/testdata/exampledata/quarkus-deps-cyclonedx.json
