+++ internal/testing/testdata/exampledata/small-deps-cyclonedx.json
