+++ internal/testing/testdata/exampledata/small-spdx.json
