+++ pkg/handler/processor/cyclonedx/cyclonedx_test.go
@@ -20,7 +20,7 @@ import (
 	"reflect"
 	"testing"
 
-	testdata "github.com/guacsec/guac/internal/testing/processor"
+	"github.com/guacsec/guac/internal/testing/testdata"
 	"github.com/guacsec/guac/pkg/handler/processor"
 )
