+++ pkg/handler/processor/process/process_test.go
@@ -17,12 +17,10 @@ package process
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
-	"reflect"
 	"testing"
 
-	"github.com/guacsec/guac/internal/testing/ingestor/simpledoc"
+	"github.com/guacsec/guac/internal/testing/dochelper"
+	"github.com/guacsec/guac/internal/testing/simpledoc"
 	"github.com/guacsec/guac/pkg/handler/processor"
 	"github.com/guacsec/guac/pkg/handler/processor/guesser"
 )
@@ -45,7 +43,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 			Format:            processor.FormatJSON,
 			SourceInformation: processor.SourceInformation{},
 		},
-		expected: docNode(&processor.Document{
+		expected: dochelper.DocNode(&processor.Document{
 			Blob: []byte(`{
 						"issuer": "google.com",
 						"info": "this is a cool document"
@@ -73,7 +71,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 			Format:            processor.FormatJSON,
 			SourceInformation: processor.SourceInformation{},
 		},
-		expected: docNode(
+		expected: dochelper.DocNode(
 			&processor.Document{ //root
 				Blob: []byte(`{
                            "issuer": "google.com",
@@ -90,7 +88,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 				Format:            processor.FormatJSON,
 				SourceInformation: processor.SourceInformation{},
 			},
-			docNode(&processor.Document{ // child 1
+			dochelper.DocNode(&processor.Document{ // child 1
 				Blob: []byte(`{
 						"issuer": "google.com",
 						"info": "this is a cooler nested doc 1"
@@ -99,7 +97,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 				Format:            processor.FormatJSON,
 				SourceInformation: processor.SourceInformation{},
 			}),
-			docNode(&processor.Document{ //child 2
+			dochelper.DocNode(&processor.Document{ //child 2
 				Blob: []byte(`{
 						"issuer": "google.com",
 						"info": "this is a cooler nested doc 2"
@@ -135,7 +133,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 			Format:            processor.FormatJSON,
 			SourceInformation: processor.SourceInformation{},
 		},
-		expected: docNode(
+		expected: dochelper.DocNode(
 			&processor.Document{ // root
 				Blob: []byte(`{
                            "issuer": "google.com",
@@ -160,7 +158,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 				Format:            processor.FormatJSON,
 				SourceInformation: processor.SourceInformation{},
 			},
-			docNode(&processor.Document{
+			dochelper.DocNode(&processor.Document{
 				Blob: []byte(`{
 							"issuer": "google.com",
 							 "info": "this is a cooler nested doc 1",
@@ -173,7 +171,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 				Format:            processor.FormatJSON,
 				SourceInformation: processor.SourceInformation{},
 			},
-				docNode(&processor.Document{
+				dochelper.DocNode(&processor.Document{
 					Blob: []byte(`{
                         "issuer": "google.com",
                         "info": "this is a cooler nested doc 3"
@@ -182,7 +180,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 					Format:            processor.FormatJSON,
 					SourceInformation: processor.SourceInformation{},
 				})),
-			docNode(&processor.Document{
+			dochelper.DocNode(&processor.Document{
 				Blob: []byte(`{
                               "issuer": "google.com",
                                "info": "this is a cooler nested doc 2",
@@ -195,7 +193,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 				Format:            processor.FormatJSON,
 				SourceInformation: processor.SourceInformation{},
 			},
-				docNode(&processor.Document{
+				dochelper.DocNode(&processor.Document{
 					Blob: []byte(`{
                         "issuer": "google.com",
                         "info": "this is a cooler nested doc 4"
@@ -227,7 +225,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 			Format:            processor.FormatJSON,
 			SourceInformation: processor.SourceInformation{},
 		},
-		expected: docNode(
+		expected: dochelper.DocNode(
 			&processor.Document{ //root
 				Blob: []byte(`{
                    "issuer": "google.com",
@@ -248,7 +246,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 				Format:            processor.FormatJSON,
 				SourceInformation: processor.SourceInformation{},
 			},
-			docNode(&processor.Document{ // child 1
+			dochelper.DocNode(&processor.Document{ // child 1
 				Blob: []byte(`{
 				"issuer": "google.com",
 				"info": "this is a cooler nested doc 1"
@@ -257,7 +255,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 				Format:            processor.FormatJSON,
 				SourceInformation: processor.SourceInformation{},
 			}),
-			docNode(&processor.Document{ // child 2
+			dochelper.DocNode(&processor.Document{ // child 2
 				Blob: []byte(`{
 				"issuer": "google.com",
 				"info": "this is a cooler nested doc 2",
@@ -270,7 +268,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 				Format:            processor.FormatJSON,
 				SourceInformation: processor.SourceInformation{},
 			},
-				docNode(&processor.Document{ // child 2.1
+				dochelper.DocNode(&processor.Document{ // child 2.1
 					Blob: []byte(`{
                   "issuer": "google.com",
                   "info": "this is a cooler nested doc 4"
@@ -373,7 +371,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 				Source:    "a-source",
 			},
 		},
-		expected: docNode(&processor.Document{
+		expected: dochelper.DocNode(&processor.Document{
 			Blob: []byte(`{
 						"issuer": "google.com",
 						"info": "this is a cool document"
@@ -408,7 +406,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 				Source:    "a-source",
 			},
 		},
-		expected: docNode(
+		expected: dochelper.DocNode(
 			&processor.Document{ //root
 				Blob: []byte(`{
                            "issuer": "google.com",
@@ -428,7 +426,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 					Source:    "a-source",
 				},
 			},
-			docNode(&processor.Document{ //child 1
+			dochelper.DocNode(&processor.Document{ //child 1
 				Blob: []byte(`{
                         "issuer": "google.com",
                         "info": "this is a cooler nested doc 1"
@@ -440,7 +438,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 					Source:    "a-source",
 				},
 			}),
-			docNode(&processor.Document{ //child 2
+			dochelper.DocNode(&processor.Document{ //child 2
 				Blob: []byte(`{
                         "issuer": "google.com",
                         "info": "this is a cooler nested doc 2"
@@ -465,7 +463,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 			Format:            processor.FormatUnknown,
 			SourceInformation: processor.SourceInformation{},
 		},
-		expected: docNode(&processor.Document{
+		expected: dochelper.DocNode(&processor.Document{
 			Blob: []byte(`{
 						"issuer": "google.com",
 						"info": "this is a cool document"
@@ -486,7 +484,7 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 			Format:            processor.FormatUnknown,
 			SourceInformation: processor.SourceInformation{},
 		},
-		expected: docNode(&processor.Document{
+		expected: dochelper.DocNode(&processor.Document{
 			Blob: []byte(`{
 						"issuer": "google.com",
 						"info": "this is a cool document"
@@ -519,8 +517,8 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 				return
 			}
 
-			if !docTreeEqual(docTree, tt.expected) {
-				t.Errorf("doc tree did not match up, got\n%s, \nexpected\n%s", stringTree(docTree), stringTree(tt.expected))
+			if !dochelper.DocTreeEqual(docTree, tt.expected) {
+				t.Errorf("doc tree did not match up, got\n%s, \nexpected\n%s", dochelper.StringTree(docTree), dochelper.StringTree(tt.expected))
 			}
 
 			/*
@@ -538,91 +536,4 @@ func Test_SimpleDocProcessTest(t *testing.T) {
 			*/
 		})
 	}
-}
-
-func existAndPop(nodes []*processor.DocumentNode, n *processor.DocumentNode) bool {
-	for i, nn := range nodes {
-		if docNodeEqual(nn, n) {
-			nodes = append(nodes[:i], nodes[i+1:]...) //nolint: staticcheck
-			return true
-		}
-	}
-	return false
-}
-
-func docEqual(a, b *processor.Document) bool {
-	a.Blob = consistentJsonBytes(a.Blob)
-	b.Blob = consistentJsonBytes(b.Blob)
-	return reflect.DeepEqual(a, b)
-}
-
-func docTreeEqual(a, b processor.DocumentTree) bool {
-	return docNodeEqual(a, b)
-}
-
-func docNodeEqual(a, b *processor.DocumentNode) bool {
-	if a == nil || b == nil {
-		return false
-	}
-
-	// check if a and b Docuemnts are equal
-	if !docEqual(a.Document, b.Document) {
-		return false
-	}
-
-	// check if len of children are equal
-	if len(a.Children) != len(b.Children) {
-		return false
-	}
-
-	if len(a.Children) > 0 {
-		// Copy list of documentNodes of A
-		aCopy := make([]*processor.DocumentNode, len(a.Children))
-		copy(aCopy, a.Children)
-
-		// for each document in B, check exists and pop on listA
-		// where exists and pop equivalency
-		for _, bNode := range b.Children {
-			if !existAndPop(aCopy, bNode) {
-				return false
-			}
-		}
-	}
-
-	return true
-}
-
-// consistentJsonBytes makes sure that the blob byte comparison
-// does not differ due to whitespace in testing definitions.
-func consistentJsonBytes(b []byte) []byte {
-	var v interface{}
-	err := json.Unmarshal(b, &v)
-	if err != nil {
-		panic(err)
-	}
-	out, _ := json.Marshal(v)
-	return out
-}
-
-func docNode(v *processor.Document, children ...*processor.DocumentNode) *processor.DocumentNode {
-	return &processor.DocumentNode{
-		Document: v,
-		Children: children,
-	}
-}
-
-func stringTree(n *processor.DocumentNode) string {
-	return stringTreeHelper(n, "")
-}
-
-func stringTreeHelper(n *processor.DocumentNode, prefix string) string {
-	str := fmt.Sprintf("%s { doc: %s, %v, %v, %v}", prefix, string(consistentJsonBytes(n.Document.Blob)),
-		n.Document.Format,
-		n.Document.Type,
-		n.Document.SourceInformation,
-	)
-	for _, c := range n.Children {
-		str += "\n" + stringTreeHelper(c, prefix+"-")
-	}
-	return str
 }
